+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - support non-standart feed filenames
 
+### Fixed
+- websub pinger would panic on response error
+
 ## [0.2.0] - 2019-05-06
 ### Added
 - support for multiple WebSub hubs